+++ lib/krad_compositor/krad_text.h
@@ -55,8 +55,8 @@ void krad_text_set_font (krad_text_t *krad_text, char *font);
 void krad_text_set_new_scale (krad_text_t *krad_text, float scale);
 void krad_text_set_new_opacity (krad_text_t *krad_text, float opacity);
 void krad_text_set_new_rotation (krad_text_t *krad_text, float rotation);
-void krad_text_set_rgb (krad_text_t *krad_text, int red, int green, int blue);
-void krad_text_set_new_rgb (krad_text_t *krad_text, int red, int green, int blue);
+void krad_text_set_rgb (krad_text_t *krad_text, float red, float green, float blue);
+void krad_text_set_new_rgb (krad_text_t *krad_text, float red, float green, float blue);
 
 void krad_text_set_tickrate (krad_text_t *krad_text, int tickrate);
 void krad_text_render (krad_text_t *krad_text, cairo_t *cr);
